requirements.txt
torch>=1.12.0
torch-geometric>=2.0.0
numpy>=1.21.0
scipy>=1.7.0
pyyaml>=5.4
tqdm>=4.60
matplotlib>=3.4
scikit-learn>=1.0
pandas>=1.3.0
